refactored DbManager and Services in Backend
