+++ Backend/src/router/authRouter.ts
@@ -1,12 +1,11 @@
 import { Router } from "express";
-import { registerUser, loginUser } from "../controller/authController";
-
+import AuthController from "../controller/AuthController";
 const authRouter = Router();
 
 // Route for user registration
-authRouter.post("/register", registerUser);
+authRouter.post("/register",AuthController.registerUser);
 
 // Route for user login
-authRouter.post("/login", loginUser);
+authRouter.post("/login", AuthController.loginUser);
 
 export default authRouter;
