+++ Backend/src/router/todosRouter.ts
@@ -1,20 +1,14 @@
 import { Router } from "express";
 import { verifyToken } from "../middleware/verifyToken";
-import {
-  addTodo,
-  deleteTodo,
-  updateTodo,
-  getTodos,
-  toggleComplete,
-} from "../controller/todosController";
+import TodoController from "../controller/TodosController";
 
 const todosRouter = Router();
 
 // Todo Routes
-todosRouter.post("/todos", verifyToken, addTodo);
-todosRouter.get("/todos", verifyToken, getTodos);
-todosRouter.put("/todos/:id", verifyToken, updateTodo);
-todosRouter.delete("/todos/:id", verifyToken, deleteTodo);
-todosRouter.put("/todos/toggle/:id", verifyToken, toggleComplete);
+todosRouter.post("/todos", verifyToken, TodoController.addTodo);
+todosRouter.get("/todos", verifyToken,TodoController.getTodos);
+todosRouter.put("/todos/:id", verifyToken, TodoController.updateTodo);
+todosRouter.delete("/todos/:id", verifyToken, TodoController.deleteTodo);
+todosRouter.put("/todos/toggle/:id", verifyToken, TodoController.toggleComplete);
 
 export default todosRouter;
