--- Backend/src/services/GetTodoServices.ts
@@ -1,12 +1,0 @@
-import { GetTodoManager} from "../DbManager/GetTodoManager"; // Adjust the path
-
-class GetTodoService {
-  static async getTodos(userId: number): Promise<object[]> {
-    // Call the DB manager to fetch todos
-    const todos = await GetTodoManager.getTodosByUserId(userId);
-
-    return todos;
-  }
-}
-
-export { GetTodoService };
